User Role modifing ...

+++ 05_project/app/src/main/java/project/tripMaker/dao/ScheduleDao.java
@@ -35,5 +35,5 @@ public interface ScheduleDao {
 
   Thema getThema(int themaNo);
 
-  List<Trip> findTripsByUserNo(int userNo);
+  List<Trip> findTripsByUserNo(Long userNo);
 }
