+++ 05_project/app/src/main/java/project/tripMaker/vo/Comment.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class Comment {
     private int commentNo;             // 댓글번호
     private int boardNo;               // 게시판번호
-    private int userNo;                // 유저번호
+    private Long userNo;                // 유저번호
     private String commentContent;     // 댓글내용
     private Date commentCreatedDate;   // 댓글작성시간
 }
